% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_time.R
\name{delayed_assign_example}
\alias{delayed_assign_example}
\title{Demonstrate the effects of "Delayed assign"}
\usage{
delayed_assign_example()
}
\value{
the current time
}
\description{
When the package is loaded, this function will be evaulated and assigned to
a variable called \code{LOAD_TIME}.
}
\examples{
LOAD_TIME
delayed_assign_example()
}
