% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{set_this_lesson}
\alias{set_this_lesson}
\alias{clear_this_lesson}
\alias{this_lesson}
\title{Internal cache for storing lesson objects}
\usage{
set_this_lesson(path)

clear_this_lesson()

this_lesson(path)
}
\arguments{
\item{path}{a path to the current lesson}
}
\description{
a test function
}
\details{
details
}
\keyword{internal}
